requirements.txt
openai
python-dotenv
sentence-transformers
faiss-cpu